linting and consolidated fisher tests to one file
